requirements.txt
celery==5.5.3
Django==4.2.22
django-celery-beat==2.8.1
djangorestframework==3.16.0
pytz==2025.2
redis==6.2.0